requirements.txt
gspread==6.0.0
google-auth==2.35.0
psycopg[binary]==3.2.1
PyYAML==6.0.2
python-dotenv==1.0.1
